remove resolver
