+++ Cargo.toml
@@ -34,9 +34,7 @@ serde_json = "1.0.85"
 url = { version = "2.3.1" }
 import_map = { version = "0.21.0", features = ["ext"], optional = true }
 thiserror = "2"
-dashmap = "5.5.3"
 deno_error = { version = "0.5.3", features = ["url"] }
-deno_media_type = { version = "0.2.5", features = ["url"] }
 boxed_error = "0.2.3"
 deno_semver = { version = "0.7.1", optional = true }
 deno_package_json = { version = "0.4.0", optional = true }
